requirements.txt
numpy>=1.24
matplotlib>=3.5
yfinance>=0.2
pandas>=1.5
scipy>=1.11
fastapi>=0.110
uvicorn>=0.29
websocket-client>=1.8
openpyxl>=3.1
python-docx>=1.1
pdfminer.six>=20231228
python-multipart>=0.0.9
